Remove IssueInput and move selectIssue component in summary form  - completed summar pillar section
