--- app/views/EditAry/AssessmentRegistryForm/SummaryForm/PillarItem/SubPillarItem/IssueInput/styles.css
@@ -1,4 +1,0 @@
-.issue-input {
-    display: grid;
-    gap: var(--dui-spacing-extra-large);
-}
