+++ app/views/EditAry/AssessmentRegistryForm/SummaryForm/PillarItem/SubPillarItem/index.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import { Header, Modal, QuickActionButton, useModalState } from '@the-deep/deep-ui';
 import { IoAddCircleOutline } from 'react-icons/io5';
 import {
-    EntriesAsList,
+    EntriesAsList, Error,
 } from '@togglecorp/toggle-form';
 
 import {
@@ -11,7 +11,7 @@ import {
     PartialFormType,
 } from '#views/EditAry/AssessmentRegistryForm/formSchema';
 
-import IssueInput from './IssueInput';
+import SelectIssueInput from './SelectIssueInput';
 import AddIssueModal from '../../AddIssueModal';
 import { PillarType } from '../..';
 
@@ -24,6 +24,7 @@ export interface Props {
     disabled?: boolean;
     value: SubPillarIssueType[] | undefined;
     onChange: (...entries: EntriesAsList<PartialFormType>) => void;
+    error: Error<PartialFormType['summarySubPillarIssue']>;
 
     issuesOptions?: SummaryIssueType[] | null;
     setIssuesOptions: React.Dispatch<React.SetStateAction<SummaryIssueType[] |undefined | null>>;
@@ -40,6 +41,7 @@ function SubPillarItem(props: Props) {
         setIssuesOptions,
         value,
         onChange,
+        error,
         issueItemToClientIdMap,
         setIssueItemToClientIdMap,
     } = props;
@@ -49,6 +51,28 @@ function SubPillarItem(props: Props) {
         showModal,
         closeModal,
     ] = useModalState(false);
+
+    const getFieldValue = useCallback(
+        (n: string) => {
+            const clientId = issueItemToClientIdMap[n];
+            if (!clientId) {
+                return undefined;
+            }
+            const itemInValue = value?.find((item) => item.clientId === clientId);
+            return itemInValue;
+        }, [value, issueItemToClientIdMap],
+    );
+
+    const getMainIndex = useCallback(
+        (n: string) => {
+            const clientId = issueItemToClientIdMap[n];
+            if (!clientId) {
+                return undefined;
+            }
+            const mainIndex = value?.findIndex((item) => item.clientId === clientId);
+            return mainIndex;
+        }, [value, issueItemToClientIdMap],
+    );
 
     return (
         <div className={styles.subPillarItem}>
@@ -65,17 +89,105 @@ function SubPillarItem(props: Props) {
                     </QuickActionButton>
                 )}
             />
-            <IssueInput
-                name={name}
-                subPillar={data.subPillar}
-                value={value}
-                onChange={onChange}
-                issueOptions={issuesOptions}
-                setIssueOptions={setIssuesOptions}
-                disabled={disabled}
-                issueItemToClientIdMap={issueItemToClientIdMap}
-                setIssueItemToClientIdMap={setIssueItemToClientIdMap}
-            />
+            <div className={styles.issueInput}>
+                <SelectIssueInput
+                    name={`${name}-1`}
+                    order={1}
+                    placeholder="1. Field Name"
+                    value={getFieldValue(`${name}-1`)}
+                    mainIndex={getMainIndex(`${name}-1`)}
+                    onChange={onChange}
+                    issuesOptions={issuesOptions}
+                    setIssueItemToClientIdMap={setIssueItemToClientIdMap}
+                    setIssuesOptions={setIssuesOptions}
+                    subPillar={data.subPillar}
+                    disabled={disabled}
+                    error={error}
+                />
+                <SelectIssueInput
+                    name={`${name}-2`}
+                    order={2}
+                    placeholder="2. Field Name"
+                    value={getFieldValue(`${name}-2`)}
+                    mainIndex={getMainIndex(`${name}-2`)}
+                    onChange={onChange}
+                    issuesOptions={issuesOptions}
+                    setIssueItemToClientIdMap={setIssueItemToClientIdMap}
+                    setIssuesOptions={setIssuesOptions}
+                    subPillar={data.subPillar}
+                    disabled={disabled}
+                    error={error}
+                />
+                <SelectIssueInput
+                    name={`${name}-3`}
+                    order={3}
+                    placeholder="3. Field Name"
+                    value={getFieldValue(`${name}-3`)}
+                    mainIndex={getMainIndex(`${name}-3`)}
+                    onChange={onChange}
+                    issuesOptions={issuesOptions}
+                    setIssueItemToClientIdMap={setIssueItemToClientIdMap}
+                    subPillar={data.subPillar}
+                    setIssuesOptions={setIssuesOptions}
+                    disabled={disabled}
+                    error={error}
+                />
+                <SelectIssueInput
+                    name={`${name}-4`}
+                    order={4}
+                    placeholder="4. Field Name"
+                    value={getFieldValue(`${name}-4`)}
+                    mainIndex={getMainIndex(`${name}-4`)}
+                    onChange={onChange}
+                    issuesOptions={issuesOptions}
+                    setIssueItemToClientIdMap={setIssueItemToClientIdMap}
+                    subPillar={data.subPillar}
+                    setIssuesOptions={setIssuesOptions}
+                    disabled={disabled}
+                    error={error}
+                />
+                <SelectIssueInput
+                    name={`${name}-5`}
+                    order={5}
+                    placeholder="5. Field Name"
+                    value={getFieldValue(`${name}-5`)}
+                    mainIndex={getMainIndex(`${name}-5`)}
+                    onChange={onChange}
+                    issuesOptions={issuesOptions}
+                    setIssueItemToClientIdMap={setIssueItemToClientIdMap}
+                    subPillar={data.subPillar}
+                    setIssuesOptions={setIssuesOptions}
+                    disabled={disabled}
+                    error={error}
+                />
+                <SelectIssueInput
+                    name={`${name}-6`}
+                    order={6}
+                    placeholder="6. Field Name"
+                    value={getFieldValue(`${name}-6`)}
+                    mainIndex={getMainIndex(`${name}-6`)}
+                    onChange={onChange}
+                    issuesOptions={issuesOptions}
+                    setIssueItemToClientIdMap={setIssueItemToClientIdMap}
+                    subPillar={data.subPillar}
+                    setIssuesOptions={setIssuesOptions}
+                    disabled={disabled}
+                    error={error}
+                />
+                <SelectIssueInput
+                    name={`${name}-7`}
+                    order={7}
+                    placeholder="7. Field Name"
+                    value={getFieldValue(`${name}-7`)}
+                    mainIndex={getMainIndex(`${name}-7`)}
+                    onChange={onChange}
+                    issuesOptions={issuesOptions}
+                    setIssueItemToClientIdMap={setIssueItemToClientIdMap}
+                    subPillar={data.subPillar}
+                    setIssuesOptions={setIssuesOptions}
+                    error={error}
+                />
+            </div>
             {isModalShown && (
                 <Modal
                     heading="Issue Editor"
